stub out some gameplay pieces like Army and Tile
