+++ src/BackGroundCanvas.ts
@@ -1,4 +1,5 @@
 import { Map } from './Map';
+import { Tile } from './Tile';
 
 export class BackgroundMap {
     private canvas : HTMLCanvasElement;
@@ -43,7 +44,7 @@ export class BackgroundMap {
 
     public SelectTile(x : number, y : number) {
         if (this.selected) {
-            this.selected.undraw(this.getTileWidth(),this.getTileHeight(),this.ctx);
+            this.selected.clear(this.getTileWidth(),this.getTileHeight(),this.ctx);
         }
 
         this.selected = new Tile(
@@ -57,25 +58,4 @@ class MapView {
     public XPos : number;
     public YPos : number;
     public scale : number;
-}
-
-
-class Tile {
-    readonly x : number;
-    readonly y : number;
-    constructor(x : number, y : number) {
-        this.x = x;
-        this.y = y;
-    }
-
-    public undraw(width : number, height : number, ctx : CanvasRenderingContext2D) {
-        ctx.clearRect(width*this.x+1, height*this.y+1, width - 2, height -2);
-    }
-
-    public draw(width : number, height : number, ctx : CanvasRenderingContext2D) {
-        ctx.beginPath();
-        ctx.rect(width*this.x + 5, height*this.y+5, width - 10, height -10);
-        ctx.rect
-        ctx.stroke();
-    }
 }
