+++ src/Tile.ts
@@ -1,0 +1,27 @@
+import { Army } from './Army';
+
+export class Tile {
+    readonly x : number;
+    readonly y : number;
+    armies : Army[];
+
+    constructor(x : number, y : number) {
+        this.x = x;
+        this.y = y;
+    }
+
+    public registerArmy(army : Army) {
+        this.armies.push(army);
+    }
+
+    public clear(width : number, height : number, ctx : CanvasRenderingContext2D) {
+        ctx.clearRect(width*this.x+1, height*this.y+1, width - 2, height -2);
+    }
+
+    public draw(width : number, height : number, ctx : CanvasRenderingContext2D) {
+        ctx.beginPath();
+        ctx.rect(width*this.x + 5, height*this.y+5, width - 10, height -10);
+        ctx.rect
+        ctx.stroke();
+    }
+}
