+++ src/main.ts
@@ -8,10 +8,10 @@ export function runMain() {
 var bgm: BackgroundMap;
 export function initCanvas(canvas: HTMLCanvasElement) {
   bgm = new BackgroundMap({ canvas: canvas, map: { width: 10, height: 5 } });
+  bgm.drawMap();
 }
 
 export function animate() {
-  bgm.drawMap();
 }
 
 export function onGameClick(x: number, y: number) {
